refactor: move image preview to CampaignInfoStep and restructure layout
